Master TeX Preamble
This is an academic document template created without any additional dependencies or technical requirements. Please note that this is version = v1.0 officially released on 28 Mar 2025, written by Shubhajit Dey (shubhajitdey412@icloud.com). It may be distributed and/or modified under the conditions of the LaTeX Project Public License, either version 1.3 of this license or (at your option) any later version. The latest version of this license is in http://www.latex-project.org/lppl.txt and version 1.3 or later is part of all distributions of LaTeX version 2003/12/01 or later.